+++ lib/src/main/java/org/eclipse/tsp/java/client/api/xy/dto/GetXyModelRequestDto.java
@@ -22,5 +22,5 @@ public class GetXyModelRequestDto {
 
 	@JsonProperty("requested_items")
 	@NonNull
-	private List<Integer> requestedItems;
+	private List<Long> requestedItems;
 }
